+++ modules/utils.py
@@ -42,7 +42,7 @@ def save_gif(frames):
 
         agent_quantity = config['agents']['quantity']
         task_quantity = config['tasks']['quantity']
-        decision_making_module_path = config['agents']['decision_making_module_path']
+        decision_making_module_path = config['decision_making']['plugin']
         module_path, class_name = decision_making_module_path.rsplit('.', 1)
         datetime_now = datetime.datetime.now()
         current_time_string = datetime_now.strftime("%Y-%m-%d_%H-%M-%S")